agregando archivo SQL y mejorando métodos

+++ asisbiom_backend/src/main/java/eetp612/com/ar/asisbiom/horarios/Horario.java
@@ -55,6 +55,6 @@ public class Horario {
     private Dia dia;
     
     @Column(name = "valor_inasistencia")
-    private Integer valorInasistencia;
+    private float valorInasistencia;
 
 }
